requirements.txt
python-telegram-bot
beautifulsoup4
requests
flask